--- a/go.mod
+++ b/go.mod
@@ -4,6 +4,7 @@ go 1.15
 
 require (
 	github.com/go-redis/redis/v8 v8.8.0
+	github.com/segmentio/kafka-go v0.4.12 // indirect
 	github.com/stretchr/testify v1.7.0
 	github.com/tidwall/btree v0.4.2 // indirect
 	github.com/tidwall/buntdb v1.2.0