This change implements a hierarchical timerwheel, and uses kafka to persist events. The `timert` interface is currently quite synchronous, and as a result, the timer returns from most actions before ensuring that the changes are persisted to kafka. If we wish to be more safe around persistence (ie return a promise or accept callback to check for errors in persistance), we'll need to introduce a more asynchronous interface, and scale writers/partitions to maximize throughput, but I'd consider the "unsafe" results collected from a test run on my laptop below to be an achievable goal in a tuned "safe" implementation:

persisting to kafka (kafka:9092 perf-20210407134512)...
Create 1000000 timer used 22.1465828s
Cancel 1000000 timer used 9.0456587s
Current time: 1617803112
Total created: 1000000, expired: 7173, canceled: 992827
Average tick process time: 66.571µs (n: 1326, min: 5.8µs, max: 7.6804ms)
Now if there's still timer:
Current time: 1617803112
Total created: 1000000, expired: 7173, canceled: 992827
Average tick process time: 47.365µs (n: 1953, min: 5.8µs, max: 7.6804ms)