@shua 2021-04-07
I was running it in our event-processor docker env using
```
docker run -it --rm -v $PWD:/go -e GOPATH=/gopath --network jep_default golang go run . 
```
to ensure kafka was available at `kafka:9092`, but I can also change it to find kafka addr via env var